Refactor to use ktest_context
Replace dependencies on (ib)devices and the old test_dev
with a new generic ktest_context.

Add support for 0 or more context names to run a test in.

Added selftest.ko with a simple test of ktest_map
that caused me to find two bugs.. :-)

The query function in this version concatenates the name
of the test with _contextname and presents one test
per context to user space. This is a trial solution,
will propagate info on handles with contexts to user space
and handle the expansion there instead.

Signed-off-by: Knut Omang <knut.omang@oracle.com>