Implement proper handling of multiple contexts
A context is now implicitly being added as _<context>
at the end of the test names, and multiple contexts will
expand to multiple test invocations with each of these
contexts.

This would allow test runs in environments where the test
context is tied to a device, and there are multiple devices
that should be tested with the same tests.

Signed-off-by: Knut Omang <knut.omang@oracle.com>