Generate symbol resolving code from symbol list
Add a new 'resolve' script which takes an input file
with a list of symbols (+ some 'directives') and
automatically generates a header file to simplify
the resolving of internal module/kernel symbols
that the test code needs to access.

The AM_KTEST_DIR m4 macro for configure.ac
supports generating rules for using the file ktest_syms.txt
in the source tree (if it exists) to generate a
ktest_syms.h file in the build tree for inclusion by
the test program.

Use ktest_gen.mk as file name for the Makefile
include file to avoid conflict with other uses

Signed-off-by: Knut Omang <knut.omang@oracle.com>