reference counting for ktf maps, test cases and tests
use kref to add refcount to ktf map elements.  In doing so interface change
is needed to allow us to pass in a "free" function for when refcount reaches
0.  For consumers who use ktf maps with dynamically-allocated content this
is convenient, and statically-defined map elements can simply pass in NULL

Rework implementation to make global test case list a kmap so it can be
refcounted using the kmap method, and similarly make a test case's list
of tests a kmap so it can use the refcount method also.  Get rid of
per-handle list of tests - it can be derived by (more expensively) walking
the set of test cases and finding those with tests that match handle.

Debugfs representation takes reference counts for the testset and the test
since they can be dereferenced from the inode.  And of course the
map takes a reference when the test case/test is added.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Change-Id: Ia65e473f4b122dc069ce3ffa371ce2485ce42df1