KTF coverage analytics
First cut at coverage analysis via kprobes.  In this patch we add
new commands to netlink protocol to enable/disable coverage on a
per-module basis.  We also add self-test around coverage analysis
but do not yet add a user-space client to enable/disable coverage.
Next step is to do precisely that, but we have enough to start with.

Coverage is done by using kallsyms to walk module and enable probes
for each function.  When the probe fires we count the number of
times a specific function is called (stored in "count" in the map
entry for the symbol, and we also count how many unique functions
in the module have been called. We can compare this to the total
number of probed functions to get a sense of overall coverage.

Coverage data can be examined via /sys/kernel/debug/ktf/coverage
debugfs file, e.g. after running selftest which enables/disables
coverage around call to the cov_counted() function, we see:

cat /sys/kernel/debug/ktf/coverage
                  MODULE               #FUNCTIONS    #CALLED
                selftest                       14          1

                  MODULE                 FUNCTION      COUNT
                selftest              myelem_free          0
                selftest               myelem_cmp          0
                selftest        ktf_return_printk          0
                selftest              cov_counted          1
                selftest                    dummy          0
                selftest             wrongversion          0
                selftest                simplemap          0
                selftest               probeentry          0
                selftest         ktf_entry_printk          0
                selftest              probereturn          0
                selftest                cov_count          0
                selftest          resolve_symbols          0
                selftest                  to_mctx          0
                selftest            selftest_exit          0

Note that coverage entries etc are global and not defined relative
to a specific handle - this is because the probe enablings themselves
are global. We refcount cases where multiple callers enable probes
and when last disable has occurred we disable them. Coverage entries
etc are only cleaned up on ktf module unload.

Note that kprobes have some limitations, particularly around inlined
functions - probes will likely not fire.  Investigating this further
to see if there's a way to figure out all inlined sites.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Change-Id: I6d01d0fbde06166b3259837038ed9c6a183c4d91