ktf coverage kprobe handler can be made safer
In some cases when enabling ktf coverage, we run into problems.  These stem
from the fact that the kprobe handler for coverage events does some risky
things like attempting to acquire a spinlock for ktf map lookup.  A much
simpler approach would be to use the pointer to the kprobe that fired to
dereference the struct ktf_cov_entry *.  If we reshuffle the cov entry
data structure to ensure the kprobe is first, and add a magic number to
ensure we only dereference entries from kprobes, this becomes an O(1) operation
rather than incurring the cost of rbtree lookup.  In some cases with the
prior approach panics resulted - the reason appearead to be that the
inlining of spinlock functions meant that if we tried to acquire a spinlock
in probe context that would trigger a recursive probe enter.

Change-Id: I440972215d2695851b14063225a8bfcb79901a35
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>