Add initializer for ktf_map and use it to fix missing inits
Some of the newly introduced struct elements in ktf_map were not
always initialized. A.o. the introduced spinlock were not initialized
for contexts, causing spinlocks to be used with random values in it.

Added __KTF_MAP_INITIALIZER to allow a common initialization of ktf_map
objects and use this where ktf_map objects are initialized inside ktf.

Change-Id: I994a4c2535f23a53bc779d1d6746fe9cb2928f29
Signed-off-by: Knut Omang <knut.omang@oracle.com>