fix panic in jprobe_return_end() triggered by selftest/probeentry
It turns out the combination of gkdb and jprobes (entry probes
abstraction in kprobes) is toxic.  We can replace our usage of
jprobes with kretprobes, since the latter also support probing
function entry.  Doing this simplifies the specification of
KTF_ENTRY_PROBE(), since we can now just pass in the name of
the function we wish to probe and its handler (the latter is
needed to allow us to support setting different handlers for
the same function, as we might need for different tests). We can
abstract the entry/return definitions since the only difference now
is whether the probe's .entry_handler or .handler function is set
(for entry and return respectively).  rdstest/rdstest_api.c will
need a follow-on fix since it uses RETURN_PROBE() APIs but only
specifies the to-be-probed function, not a handler name as the
new (and more consistent) API requires.

Also fix off-by-one error which resulted in us not enabling
coverage for adjacent functions.  We figure out if we already
have coverage for a function by looking up our ktf map using
the address + size.  This lookup was broken because with
adjacent functions f1 and f2, we would look up f2 and because
f1 + size(f1) == f2 we would assume we already had f2 covered.
This caused us to miss coverage for functions, so in fact
for lookup we need to check if f2 >= f1 + size(f1).

Change-Id: Idffc3ea5faed7f2fd3e8ddc30967c5341da70472
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>