ktf: add postmortem debugging doc
if a test run crashes the kernel it is useful to be able to debug
the crash dump to figure out what KTF was doing.  This doc describes
how we load symbols for KTF and traverse the set of test cases and
tests, examining logs and which test was running when panic ensued.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Change-Id: Iad5972d64c91ef9c34f9f085df15398f3122a5ae