fix panic, coverage issues around unloading/reloading test  modules
KTF contexts were not implicitly removed on cleaning up handles.  Worse,
handle references stick around due to the context_handles list.  Update
handle cleanup to warn about any remaining contexts prior to handle cleanup.

In testing these changes it became obvious that coverage support across
unload/reload of modules was broken.  Because a function's kprobe was
originally registered with one address and on reload it could have
another we were not seeing probe firing for coverage.  The fix is to
first register kprobes on coverage re-enable, then update the coverage
entry map with the new addresses (and possibly new size in the case of
a recompiled module).  It would of course be simpler just to remove
coverage entries for functions on disable, but often we wish to retain
coverage info after unloading a module.

Change-Id: I21123afd367629db618e580d70a63012f20a78c1
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>