Make hybrid tests' effects on unrelated runs less confusing.
If a combined test - to be renamed to hybrid tests
in a later commit - introduces a test suite name,
if that suite does not have any kernel tests loaded, it
may "eat" up execution of another later kernel test suite.

This is due to the way the test runner is relying on the number
of suites to match the kernel test query.

Avoid this by detecting such "unresolved" user land
part of hybrid tests in the wrapper table and skip
those sets if no tests exists.

Change-Id: I5f562541e97c2b82bb7bdc917aec01f7dc18681d
Signed-off-by: Knut Omang <knut.omang@oracle.com>