Changes to support a pkgconfig based gtest
gtest has recently gained support for pkgconfig, replacing the
old and not maintained, non-standard gtest-config.

This is not yet part of any gtest release, or eg. Oracle or
Redhat/Fedora packaging, but supporting both the old gtest-config
and the more standard pkgconfig is better than only supporting
the obsoleted gtest-config.

Implement this now and assume the packages will eventually
include the new pkgconfig support.

Change-Id: I7c30dc613da9ec77343159da1b9873369cdd92b9
Signed-off-by: Knut Omang <knut.omang@oracle.com>