Make ktfrun work on Ubuntu as well.
On Ubuntu there's no package support for gtest except a package that
allows source to be installed. We do not make any attempts to support that,
instead we assume googletest is cloned from git and built as a dependency.

Here when libktf is linked against ktfrun, the linker figured out
that ktf was not needed, and just skipped linking it, even though there
are initialization code in libktf that needs to be executed.
Introduce a ktf::setup() function that can be called to reference
something in the library, to enforce initialization.
Resign and call this from ktfrun to make sure this problem does
not show up again.

Also make sure configure works even if googletest has been installed
in any of the the system paths, and prefix is not specified.

Change-Id: I86a579c459f65c388886887241e706308bc394be
Signed-off-by: Knut Omang <knut.omang@oracle.com>