Fix potential sleep while atomic and avoid lock in the common case
The introduction of an irq safe spin lock to protect
assertion count and log state in commit 2db0716e3c
caused the kmalloc to be called with interrupts off
which creates a potential deadlock scenario.

The case where tests pass is the performance critical
part and we also want to minimize any effects by KTF on
ordering. Optimize for the common case by only using the lock
in the exception path, and only to protect the log.

For the assertion count we instead use an atomic
which is safe and does not affect interrupt handling.
Alternatively we could just leave it unsafe as many tests are sequential
anyway and for those that are not, it does not have to be exact,
but that's likely to create ambiguous rare confusing situations
in threaded tests so lets keep it synchronized for now.

Change-Id: Ie6518ba13c44d5364931cb98940688ebf1e7b4c5
Signed-off-by: Knut Omang <knut.omang@oracle.com>