Handle attempts to configure w/relative paths more gracefully
KTF currently does not support invoking configure with
a relative path. Fail with an error message in configure
and also update documentation to state this limitation.

Also fix support for the special relative case
of building in-source, as there might be good use cases
for doing it, although the recommended mode is to build
in a separate build tree.

Mitigates Github issue #8.

Change-Id: I3b616c547171990a7dfac07598124c1a03cc4679
Signed-off-by: Knut Omang <knut.omang@oracle.com>