Make sure the -std=c++11 flag is set for c++
Newer versions of googletest uses c++ 2011 features.
Initially blaming googletest for not configuring this correctly,
it seems it is actually pkgconfig which does not seem to support
any macro to set C++ specific flags, strangely enough.

Set the flag if it is supported by the compiler,
otherwise set the -std=c++0x pre-11 feature flag.

This means we introduce use of AX_CHECK_COMPILE_FLAG and an extra
source file that also needs to be installed by clients.

Change-Id: I561d8de432c000f5e550eeeb56cf4d548d9ff026
Signed-off-by: Knut Omang <knut.omang@oracle.com>