Logging cleanup: Removed DM - added twarn, terr, tlogs
This was triggered by checkpatch warning PREFER_PR_LEVEL.
We want to avoid explicit printk in the code, to make it
easier to change the overall logging behaviour
when/if desired.

Left a single statement list composed of printk for now
and made it an exception, mainly because it was not trivial
to replace it.

Change-Id: I0ce92edd1d61ab1e7eb36620fc3f1680f99a0f90
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/15364