Fix as many DECL sparse issues as useful
If we tight too much, we run into compiler issues
with unused symbols if tests have been commented out.
There seems to be no good way to avoid this, so
keep DECL check disabled still, but make a
comment as of why it is disabled.

Change-Id: I4e1da481971d22effc64db9fad040e0934b31548
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/15381
Pre-check: franklin <franklin@review.no.oracle.com>
Tested-by: franklin <franklin@review.no.oracle.com>