support both string truncation and opaque keys in ktf maps
by default ktf maps use strings as keys but a custom compare function can
be used and opaque data placed in the key.  At element init time we do
not however know the model the map uses, so rather than assuming string
or opaque, we allow for a 63 character string (which we will auto-terminate
ourselves) or 63 bytes of opaque data. Adjust test to verify truncation
instead of error return.

The assumption of string and use of strncpy() was causing problems for cases
where the key was a memory address ending in 00. Due to endianness the
"string" such addresses resulted in started with a null terminator so we
ended up with a 0 key rather than the real address, leading to coverage
test failures.

Change-Id: Ib51f3666842d040916a6b257c93d0e578b65bfe7
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/15404
Pre-check: franklin <franklin@review.no.oracle.com>