Fix NEEDLESS_IF checkpatch issues
Lots of destructor functions within the kernels nowadays
support being called with a null pointer, such cases are
indicated by the corresponding version of checkpatch.
However the checks have been introduced over time so it
is not necessarily the case for older kernels.
This is the case for kmem_cache_destroy which was made
null pointer safe in 4.1 - keep the except as checkpatch
would otherwise fail on the bw comp wrapper.

Change-Id: I0003252995ec673881d896bbc4f48a71549998c9
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/15410
Pre-check: franklin <franklin@review.no.oracle.com>