Added support for configurable contexts
A context can now be set up to be configurable from user space
using a context specific parameter block.

Context configuration must be enabled from the kernel
context. User programs can then configure the context,
typically as part of setting up the environment for
testing. The context configuration will persist
througout the lifetime of the test module instance,
unless reconfigured by another test run or program.

There's a new macro to declare configurable contexts
on the kernel side, KTF_CONTEXT_ADD_CFG, which takes
an extra parameter, a callback function which will be
called when the user program tries to configure a context.
It's return value must be 0 if the configuration was successful.
The macro KTF_CONTEXT_CFG_OK() can be used to check
if a context has been configured properly. Normal,
nonconfigurable contexts are always assumed to be
properly configured.

Macros KTF_CONTEXT_ADD_TO and more has been added
for the cases where a context is added to a non-default
handle, and the selftest has been updated to use these.

On the user land side, there's two new macros,
KTF_CONTEXT_CFG to explicitly configure a named
context, and KTF_CONTEXT_CFG_FOR_TEST, which
allows a context to be configured by giving a test name
which context should be configured. The second form
is for cases where the context name is not unique,
in which case a full test name will allow unique
identification of a context. Contexts with the same
name can exist with different handles, but handle IDs are
cannot be determined compile time, as they may
vary with the order and number of test modules
loaded at any time.

Added a selftest for configurable contexts.

Change-Id: I04fed273a7b7222563686caeb26ff5104da56847
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/c/ktf/+/16003
Pre-check: franklin <franklin@review.no.oracle.com>
Tested-by: franklin <franklin@review.no.oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>