ktf_netctx: A configurable context for network connection info
This logic builds on ktf_context to implement a configurable
context structure to receive connection information on a
set of remote peers. This is useful to write tests that relies
on more than one kernel being involved, such as simple
network tests.

There's no user land "library" code in this commit (except for the
ktf_netctx.h header file which defines the datastructures
to use). This is to avoid having KTF rely on too many libraries
on the user space side.

Change-Id: I7d34f114b92e69a8cc17ea7940a63729ee9b13cb
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/c/ktf/+/16095