Add support for user space generated contexts
The kernel module can allow configuration type identifiers
to be set up such that user space can generate new contexts
of that configuration type. This is useful for scenarios where
user mode has easier access to information about contexts
available for the tests to run in.

Updated the selftest example to generate a new "context3"
on the fly.

Remove the old implicit ktf initialization as a side effect,
as we now rely on ktf::setup() to be called with a configuration
function callback to make sure the context configuration is
supplied at the right point in the initialization sequence
to allow us to generate the right test objects to the
Googletest framework.

Change-Id: I511287f9b4eec03e65875814731ea534b5a7d159
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/c/ktf/+/16094