Move a few backward compatibility checks to a new header file
We want KTF to be building and able to build tests across
as many kernel versions as possible, but let the code we build reflect
the latest APIs. Move a few compatibility wrappers to a separate
source file that is only needed for older kernel builds.

This makes the code more readable and simplifies compatibility handling
for client test modules.

Change-Id: Id0ed61368df6109500e0ba117d0705222fe026f3
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/c/ktf/+/16158