Disable the memory coverage (cov) test
There's subtle, nondeterministically behaving
cleanup issues seen related to the test. During some
executions (not all), elements of the cov_mem_cache
kmem_cache are not released upon cleanup.

Not clear whether it is the attempt to test it on "ourselves"
or whether it is a more generic issue with the functionality
itself.

Disable it until we can analyse more properly.

Change-Id: Iff6a05a5edcc53eac82ee5790a6815b59ee7f57f
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/c/ktf/+/16074
Pre-check: franklin <franklin@review.no.oracle.com>
Tested-by: franklin <franklin@review.no.oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>