Fix regression in ktfcov application
When we moved to require an explicit call to ktf::setup()
the ktfcov application was not updated.
This caused ktfcov to call netlink functions without
proper initialization, yielding just a core dump
and no KTF state update.

Fix this, and also handle the response message from the
kernel in an orderly way.

Change-Id: Iff2ece47d17dbef46a0ff43b510de28ece5a453b
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/c/ktf/+/16297