Make netlink protocol more in line with modern recommendations
We have been using a single REQ cmd for all user land requests and a single
RESP cmd for the responing kernel side, and have wrapped all the individual
commands in the TYPE attribute. This is not in line with recommendations
for netlink protocols, where this is explicitly discouraged.
Restructure to "move up" the TYPE attribute to become cmd, and eliminate
the use of REQ of RESP which is not used anyway. Instead for simplicity just
let the response packets from the kernel use the same cmd IDs as the corresponding
request, and use cmd correctly on both sides to dispatch messages.

In addition the documentation of the protocol was limited,
add some basic BNF notation for the various protocol elements
and a little more explanation.

Change-Id: I65ddbfa01c5404b7f1e0de9a898adac3a40abd32
Signed-off-by: Knut Omang <knut.omang@oracle.com>
Reviewed-on: https://review.no.oracle.com:8443/c/ktf/+/16720
Tested-by: franklin <franklin@review.no.oracle.com>