lib/ktf_run: Compile beyond API change in Googletest commit 5e0cf72b7cd
Commit 5e0cf72b7cd in Googletest introduces a backward incompatible
API change in ParameterizedTestCaseInfo<T>::AddTestPattern
which gets a 4th mandatory argument of type CodeLocation.

There's no change to versioning info associated with the change so
add the m4 function AC_CHECK_CODELOC_ARG_FOR_ADDTESTPATTERN
to test for this new argument and define

  HAVE_CODELOC_FOR_ADDTESTPATTERN

to use as conditional compile test.
In the code add a dummy CodeLocation argument.
The dummy argument is ok because failing tests from the kernel
embeds the file and line number info in the message anyway,
so it cannot use the CodeLocation.

Change-Id: I5869d982f1562b0717686699d20bb4aed81ce535
Reported-by: Denis Efremov <efremov@linux.com>
Signed-off-by: Knut Omang <knuto@ifi.uio.no>