ktf_compat: remove outer wrapper define
This define was there to make it easy to
completely eliminate ktf_compat.h for kernels that do not
need any bw.comp fixes, but as the need for this commit shows
it is rather error prone as it had not been updated when
more bw.comp fixes became necessary.

Just eliminate it and let a few blank lines be compiled
even for new kernels.

Change-Id: I05522a00f49c33145a6d6699cd1fb88ed90b2d56
Reported-by: Batiste Heron <github:Mantra84>
Signed-off-by: Knut Omang <knuto@ifi.uio.no>